Changing as per the code review comments
