+++ apps/portal/extensions/stores/es/index.js
@@ -35,39 +35,30 @@ var getAsset, getAssets, addAsset, deleteAsset;
     };
 
     var obtainAuthorizedHeaderForAPICall = function () {
-        try {
-            var password = config.authConfiguration.password;
-            if (config.authConfiguration.isSecureVaultEnabled) {
-                password = resolvePassword(password);
-            }
-            var authenticate = post(config.authenticationApi, {
-                "password": password,
-                "username": config.authConfiguration.username
-            }, {}, 'json');
-            return {
-                'Cookie': "JSESSIONID=" + authenticate.data.data.sessionId + ";",
-                'Accept': 'application/json'
-            };
-        } catch (e) {
-            throw e;
+        var password = config.authConfiguration.password;
+        if (config.authConfiguration.isSecureVaultEnabled) {
+            password = resolvePassword(password);
         }
+        var authenticate = post(config.authenticationApi, {
+            "password": password,
+            "username": config.authConfiguration.username
+        }, {}, 'json');
+        return {
+            'Cookie': "JSESSIONID=" + authenticate.data.data.sessionId + ";",
+            'Accept': 'application/json'
+        };
     };
 
     var getPublishedAssets = function () {
-        try {
-            var headers = obtainAuthorizedHeaderForAPICall();
-            var assets = parse((get(config.publishedAssetApi, headers, 'application/json')).data).data;
-            var publishedAssets = [];
-            for (var i = 0; i < assets.length; i++) {
-                var asset_id = (assets[i][GADGET_ATTRIBUTES][GADGET_ID] + config.dirNameDelimiter
-                + assets[i][GADGET_ATTRIBUTES][GADGET_VERSION]).replace(/ /g, config.dirNameDelimiter);
-                publishedAssets.push(asset_id);
-            }
-            return publishedAssets;
-        } catch (e) {
-            log.error("Cannot connect to Enterprise Store");
+        var headers = obtainAuthorizedHeaderForAPICall();
+        var assets = parse((get(config.publishedAssetApi, headers, 'application/json')).data).data;
+        var publishedAssets = [];
+        for (var i = 0; i < assets.length; i++) {
+            var asset_id = (assets[i][GADGET_ATTRIBUTES][GADGET_ID] + config.dirNameDelimiter
+            + assets[i][GADGET_ATTRIBUTES][GADGET_VERSION]).replace(/ /g, config.dirNameDelimiter);
+            publishedAssets.push(asset_id);
         }
-
+        return publishedAssets;
     };
 
     getAsset = function (type, id) {
@@ -94,41 +85,34 @@ var getAsset, getAssets, addAsset, deleteAsset;
         if (type === 'layout') {
             return;
         }
-        var publishedAssets = getPublishedAssets();
-        if (publishedAssets) {
-            var ctx = utils.currentContext();
-            var parent = new File(assetsDir(ctx, type));
-            var allAssets = parent.listFiles();
-            var assets = [];
-            query = query ? new RegExp(query, 'i') : null;
-            for (var j = 0; j < publishedAssets.length; j++) {
-                allAssets.forEach(function (file) {
-                    if (publishedAssets[j] === file.getName()) {
-                        if (!file.isDirectory()) {
-                            return;
-                        }
-                        file = new File(file.getPath() + '/' + type + '.json');
-                        if (file.isExists()) {
-                            file.open('r');
-                            var asset = JSON.parse(file.readAll());
-                            if (!query) {
-                                assets.push(asset);
-                                file.close();
-                                return;
-                            }
-                            var title = asset.title || '';
-                            if (!query.test(title)) {
-                                file.close();
-                                return;
-                            }
-                            assets.push(asset);
-                            file.close();
-                        }
-                    }
-                });
+        var ctx = utils.currentContext();
+        var parent = new File(assetsDir(ctx, type));
+        var assetz = parent.listFiles();
+        var assets = [];
+        query = query ? new RegExp(query, 'i') : null;
+        assetz.forEach(function (file) {
+            if (!file.isDirectory()) {
+                return;
             }
-            return assets;
-        }
+            file = new File(file.getPath() + '/' + type + '.json');
+            if (file.isExists()) {
+                file.open('r');
+                var asset = JSON.parse(file.readAll());
+                if (!query) {
+                    assets.push(asset);
+                    file.close();
+                    return;
+                }
+                var title = asset.title || '';
+                if (!query.test(title)) {
+                    file.close();
+                    return;
+                }
+                assets.push(asset);
+                file.close();
+            }
+        });
+        return assets;
     };
 
     /**
