+++ apps/portal/resources/extensions/es/src/dsgadget/configs/portal.json
@@ -1,9 +1,9 @@
 {
   "url": "https://localhost:9444/portal",
-  "username" : "admin",
-  "password" : "admin",
-  "gadgetFileSizeLimit" : 5,
-  "delimiter" : "_",
-  "storeType" : "es",
-  "type" : "gadget"
+  "username": "admin",
+  "password": "admin",
+  "gadgetFileSizeLimit": 5,
+  "delimiter": "_",
+  "storeType": "es",
+  "type": "gadget"
 }
